wifi: Converted to 16 bit channel width in wifi-phy

+++ src/wifi/model/wifi-phy.h
@@ -636,7 +636,7 @@ public:
    *
    * \return true if the channel definition succeeded
    */
-  bool DefineChannelNumber (uint8_t channelNumber, WifiPhyStandard standard, uint16_t frequency, uint8_t channelWidth);
+  bool DefineChannelNumber (uint8_t channelNumber, WifiPhyStandard standard, uint16_t frequency, uint16_t channelWidth);
 
   /**
    * A pair of a ChannelNumber and WifiPhyStandard
@@ -645,7 +645,7 @@ public:
   /**
    * A pair of a center Frequency and a ChannelWidth
    */
-  typedef std::pair<uint16_t, uint8_t> FrequencyWidthPair;
+  typedef std::pair<uint16_t, uint16_t> FrequencyWidthPair;
 
   /**
    * Return the Channel this WifiPhy is connected to.
@@ -1624,20 +1624,19 @@ public:
   /**
    * \return the channel width
    */
-  uint8_t GetChannelWidth (void) const;
+  uint16_t GetChannelWidth (void) const;
   /**
    * \param channelwidth channel width
    */
-  virtual void SetChannelWidth (uint8_t channelwidth);
+  virtual void SetChannelWidth (uint16_t channelwidth);
   /**
    * \param channelwidth channel width (in MHz) to support
    */
-  void AddSupportedChannelWidth (uint8_t channelwidth);
+  void AddSupportedChannelWidth (uint16_t channelwidth);
   /**
    * \return a vector containing the supported channel widths, values in MHz
    */
-  std::vector<uint8_t> GetSupportedChannelWidthSet (void) const;
-
+  std::vector<uint16_t> GetSupportedChannelWidthSet (void) const;
   /**
    * Get the power of the given power level in dBm.
    * In SpectrumWifiPhy implementation, the power levels are equally spaced (in dBm).
@@ -1781,7 +1780,7 @@ private:
    * \param width The channel width to use
    * \return the channel number if found, zero if not
    */
-  uint8_t FindChannelNumberForFrequencyWidth (uint16_t frequency, uint8_t width) const;
+  uint8_t FindChannelNumberForFrequencyWidth (uint16_t frequency, uint16_t width) const;
   /**
    * Lookup frequency/width pair for channelNumber/standard pair
    * \param channelNumber The channel number to check
@@ -1939,7 +1938,7 @@ private:
   uint16_t m_channelCenterFrequency;   //!< Center frequency in MHz
   uint16_t m_initialFrequency;         //!< Store frequency until initialization
   bool m_frequencyChannelNumberInitialized; //!< Store initialization state
-  uint8_t m_channelWidth;             //!< Channel width
+  uint16_t m_channelWidth;             //!< Channel width
 
   double m_edThresholdW;          //!< Energy detection threshold in watts
   double   m_ccaMode1ThresholdW;  //!< Clear channel assessment (CCA) threshold in watts
@@ -1967,7 +1966,7 @@ private:
   typedef std::map<ChannelNumberStandardPair,FrequencyWidthPair> ChannelToFrequencyWidthMap; //!< channel to frequency width map typedef
   static ChannelToFrequencyWidthMap m_channelToFrequencyWidth; //!< the channel to frequency width map
 
-  std::vector<uint8_t> m_supportedChannelWidthSet; //!< Supported channel width
+  std::vector<uint16_t> m_supportedChannelWidthSet; //!< Supported channel width
   uint8_t              m_channelNumber;            //!< Operating channel number
   uint8_t              m_initialChannelNumber;     //!< Initial channel number
